Rename ical-generator to ics-generator
